feat(mcp-app): replace vanilla DOM renderer with React app
